LayoutTests/sputnik/Conformance/15_Native_Objects/15.1_The_Global_Object/S15.1_A1_T2-expected.txt

S15.1_A1_T2

PASS TypeError: '[object Window]' is not a constructor (evaluating 'new this()')

TEST COMPLETE
